feat: create redirect and client-side/server-side file management
